Move copied functions to `ane_transformers.py`
